tools: introduce x509 tools
We will soon need X.509 certificates with subject alternative names. They're
a bit painful to be created because it's needed to pass them inside
a configuration file. Handling temporary files in bash is possible but imho
a bit clunky.

This commit introduces python scripts that can manage all the x.509 stuff we
might need for *TESTING* this project.

The following commits will plug them in inside our testing environment.