spec: generate packageinfo files for composer and worker
For Koji builds, it would be nice to know the version of osbuild-composer
and worker that was used to build the image.

This commit explores the approach of putting the information about the package
version in two json files:

/usr/share/osbuild-composer/packageinfo.json
/usr/share/osbuild-worker/packageinfo.json

For now, I chose to have separate NVRA fields in the files.