worker/api: deprecate the arch field
From now on, the arch should be embedded in a job type, e.g:

"jobType": "osbuild:aarch"
"jobType": "osbuild-koji:s390x"

The arch field is still supported. When it's used, the API handler just
appends the specified arch after the job types - the exactly same behaviour
as we used before.

Signed-off-by: Ondřej Budai <ondrej@budai.cz>