jobqueue: generalize jobType to queueName
Soon, we want to add more job restrictions. After this change, the workers
will be able to choose a job from a queue based not only on its type but
also on more restrictions.

This is the first step to achieve it. jobqueue.Enqueue() now accepts queueName
instead of jobType. It's the responsibility of the jobqueue user to encode
jobType into queueName.

jobqueue.Dequeue() does the same thing: it now accepts queueNames instead of
jobTypes and returns queueName instead of jobType. Once again, the caller
must do encoding/decoding the queue names.

For now, the encoding/decoding is no-op and queueName == jobType.

Signed-off-by: Ondřej Budai <ondrej@budai.cz>