worker: fallback to _weldr when no accepted job owners are set
osbuild-composer installed without any configuration is still expected
to be able to run the Weldr API without any configuration. The user
should be able to just start osbuild-composer.socket and be able to build
an image.

In this case, there's no worker config and therefore the worker accepts
only jobs without an owner. Set the accepted job owners to _weldr in this
case so all jobs created by Weldr API can use this owner.

Note that this doesn't break Weldr's backward compatibility - its jobs
are still without an owner and jobs without an owner can be assigned to
any worker.

Signed-off-by: Ondřej Budai <ondrej@budai.cz>