Run RHEL depsolving for all arches as part of CI
Previously the dnf-json-tests were run as part of base tests and they
accidentally included only Fedora systems. The obvious fix of simply
putting RHEL into the list of systems that are supposed to be tested
does not work because:
a) It would not work on Fedora and I prefer not to introduce additional
branching in the test case.
b) In would not "just" work on RHEL either because the subscription we
use is only valid for the host's architecture.
c) When used together with repository overrides (our repository
snapshots) it needs to run inside of VPN.

Also dnf-json-tests are different from the other "base" tests because
they don't actually use osbuild-composer API, instead they use internal
implementation which wraps dnf-json.

For these reasons I decided to split dnf-json-tests from the base tests
which allows me to:
a) Use the overrides and don't affect the other base tests.
b) Deploy the machine in PSI, because of the VPN requirement.